chore: move table to a separate module
